[DB-33036] resultset: fetch next might return no more results
If the last result was larger than the max message size the result set
will not be marked complete when it's fetched, but the next fetch will
return no more data.  In that case we want to return None else the
result set index will be too large and throw an error.

After fetching more results check to see if we're done; if so return
None.