[154] Rewrite tests to use pynuoadmin rather than entity
In NuoDB 4.0 and above the previous NuoDB Agent administrative layer is
deprecated in favor of the NuoDB Admin layer, and in NuoDB 4.2 and above
NuoDB Agent is obsolete.

This means that the pynuodb "entity" layer is no longer useful: instead
users should use the pynuoadmin module to manage NuoDB databases.

Further, NuoDB 4.2 provides a new Vectorized Execution Engine as its
preferred SQL interface.

This commit modifies the tests for this new world.

First, it removes the entity tests altogether as it's not possible (as
of NuoDB 4.2) to use them with a database.

Second, it creates module-level operations to re-implement the
entity-based database management (create/start/stop) used by the tests
using the pynuoadmin interface.

Unfortunately pynuoadmin still requires Python 2, while pynuodb supports
both Python 2 and 3, so also re-implement the database management using
the nuocmd CLI so that the tests will run with Python 3.  This is the
implementation which is currently used.

Finally, VEE removes some support for obsolete features (such as the
"number" SQL type) which were used by the tests; remove references to
them.