Update to use NuoDB 4.2
As of NuoDB 4.0 the NuoDB Agent management layer is deprecated; as of
NuoDB 4.2 this layer is obsolete and no longer available.

The support still exists here for those using older releases, but remove
it from the README.

Simplify the Makefile and Travis configuration by taking advantage of
TLS configuration available in the nuoadmin service script.