session.send(): Allow sending of bytes, str, or bytearray
In pynuodb 2.x we only accepted str here, then we changed it to accept
only bytes.  Instead allow users to send any of bytes, str, or
bytearray.

Add a test of the different types of send.

Note, regardless of the type of message given to send(), recv() still
always returns bytes.