circle: Use the latest NuoDB release Docker image
Change the hardcoded 5.1.2 release to use "latest".

Add support for the common-config context, which is needed to obtain the
license info needed for newer instances of NuoDB.

Modify the test configuration to install a license if (a) there is not
already one available in the domain, and (b) we have one provided in the
environment (from the common-config context).