This commit represents a change in how Date, Time, and Timestamp types are handled.
When sending a Time or Timestamp to the engine.  Naive and TimeZone
based objects are supported.  They'll be converted to the connections
timezone before sending.  On receiving a Time or Timestamp from the
engine the object is created with the timezone of the connection.  At
this time the Timestamp object is converted to Naive datetime based
upon the application's local timezone.  Note, this could be different
than the connections timezone (especially after protocol TIMESTAMP
WITHOUT TIME ZONE is implemented).  The use of IANA timezone names
and tzinfo should allow us to handle daylight savings time
okay. Ideally we would not convert to a naive datetime but leave the
timezone in place for TIMESTAMP but use naive datetime with TIMESTAMP
WITHOUT TIME ZONE.  This commit does not address changes for that
protocol.

In python 2.7, pytz is used for tzinfo,  In python >3.9 zoneinfo is
used.

There are other issues that I've seen with current handling of
timestamp that I believe are also addressed but,  I don't currently
have test cases.  That is next.

For example, I have seen issues from time.mktime in DateToTicks.  For
some dates between 1/1/1 and 9999/12/31 which we should support and I
believe this does.  with the edge cases of both UTC and connection
timezones must fit within those ranges.   So a timestamp on either of
those days might have an issue.

It looks like the protocol between client and engine is to offset the
UTC time by the standard time of the timezone irregardless if the
actual timezone falls within daylight saving time or not.  Or if the
current time is in daylight saving times or not.

This implement adjust the time appropriately to take that in account
and timestamps in daylight saving time should display correctly.

I believe the builtin calendar support and switch between Julian
Calendar and Gregorian Calednar is the similar to what the TE is doing
so timestamps before 1582 October 4, should be handled correctly.