[DB-40533] Add support for new type VECTOR(<dim>,DOUBLE)
Add encode/decode support for new VECTOR datatype.

A result column of VECTOR(<dim>, DOUBLE) shows as type
datatype.VECTOR_DOUBLE in the metadata description.

It is returned as datatype.Vector with subtype
datatype.Vector.DOUBLE which can be used as a list.

For a parameter to be detected as VECTOR(<dim>,DOUBLE)
it also needs to use the datatype.Vector datatype.
Alternatively, binding a string is also possible -
which currently includes binding a list of numbers which
will then be converted to string.

This commit also adds tests for the described behavior.
Tests require NuoDB 8.0 or better to run.